+++ lib/apiScenario/postmanCollectionRunnerClient.ts
@@ -11,7 +11,7 @@ import {
   Variable,
   VariableScope,
 } from "postman-collection";
-import { xmsSkipUrlEncoding } from "../util/constants";
+import { xmsLongRunningOperation, xmsSkipUrlEncoding } from "../util/constants";
 import { JsonLoader } from "../swagger/jsonLoader";
 import {
   ApiScenarioClientRequest,
@@ -20,7 +20,6 @@ import {
   Scope,
 } from "./apiScenarioRunner";
 import {
-  AADTokenAuthentication,
   ArmTemplate,
   Authentication,
   DelayItemMetadata,
@@ -49,11 +48,17 @@ export interface PostmanCollectionRunnerClientOption {
   jsonLoader: JsonLoader;
 }
 
-interface PostmanAuthOption {
-  type: Authentication["type"];
+interface PostmanAADTokenAuthOption {
+  type: "AADToken";
   tokenName: string;
-  scriptLocation: "Collection" | "Folder" | "Request";
 }
+
+interface PostmanAzureKeyAuthOption {
+  type: "AzureKey";
+  keyName: string;
+}
+
+type PostmanAuthOption = PostmanAADTokenAuthOption | PostmanAzureKeyAuthOption;
 
 export class PostmanCollectionRunnerClient implements ApiScenarioRunnerClient {
   private opts: PostmanCollectionRunnerClientOption;
@@ -70,24 +75,56 @@ export class PostmanCollectionRunnerClient implements ApiScenarioRunnerClient {
   }
 
   private checkAuthOption(
-    auth: Authentication,
-    location: PostmanAuthOption["scriptLocation"]
-  ): PostmanAuthOption | undefined {
-    if (auth.type === "AADToken" && auth.scope) {
-      if (!this.authOptionMap.has(auth.scope)) {
-        this.authOptionMap.set(auth.scope, {
-          type: auth.type,
+    authentication: Authentication,
+    preScripts: string[],
+    env: VariableEnv
+  ): RequestAuth | undefined {
+    if (authentication.type === "AADToken") {
+      authentication.scope = env.resolveString(authentication.scope);
+      let option: PostmanAADTokenAuthOption;
+      if (!this.authOptionMap.has(authentication.scope)) {
+        option = {
+          type: authentication.type,
           tokenName: `x_bearer_token_${this.authOptionMap.size}`,
-          scriptLocation: location,
-        });
-      }
-      return this.authOptionMap.get(auth.scope);
-    } else if (auth.type === "AzureKey") {
-      // TODO
-    } else if (auth.type === "None") {
-      // TODO
-    }
+        };
+        this.authOptionMap.set(authentication.scope, option);
 
+        preScripts.push(PostmanHelper.generateAuthScript(authentication.scope, option.tokenName));
+      } else {
+        option = this.authOptionMap.get(authentication.scope)! as PostmanAADTokenAuthOption;
+      }
+      return new RequestAuth({
+        type: "bearer",
+        bearer: [
+          {
+            key: "token",
+            value: `{{${option.tokenName}}}`,
+            type: "string",
+          },
+        ],
+      });
+    } else if (authentication.type === "AzureKey") {
+      return new RequestAuth({
+        type: "apikey",
+        apikey: [
+          {
+            key: "in",
+            value: authentication.in,
+            type: "string",
+          },
+          {
+            key: "key",
+            value: authentication.name,
+            type: "string",
+          },
+          {
+            key: "value",
+            value: covertToPostmanVariable(authentication.key),
+            type: "string",
+          },
+        ],
+      });
+    }
     return undefined;
   }
 
@@ -109,25 +146,11 @@ export class PostmanCollectionRunnerClient implements ApiScenarioRunnerClient {
 
     const preScripts: string[] = [];
 
-    scenarioDef.authentication = scope.env.resolveObjectValues(scenarioDef.authentication);
-    const authOption = this.checkAuthOption(scenarioDef.authentication, "Collection");
-
-    if (authOption) {
-      this.collection.auth = new RequestAuth({
-        type: "bearer",
-        bearer: [
-          {
-            key: "token",
-            value: `{{${authOption.tokenName}}}`,
-            type: "string",
-          },
-        ],
-      });
-      preScripts.push(
-        PostmanHelper.generateAuthScript(
-          (scenarioDef.authentication as AADTokenAuthentication).scope!,
-          authOption.tokenName
-        )
+    if (scenarioDef.authentication) {
+      this.collection.auth = this.checkAuthOption(
+        scenarioDef.authentication,
+        preScripts,
+        scope.env
       );
     }
 
@@ -190,26 +213,8 @@ export class PostmanCollectionRunnerClient implements ApiScenarioRunnerClient {
 
     const preScripts: string[] = [];
 
-    scenario.authentication = env.resolveObjectValues(scenario.authentication);
-    const authOption = this.checkAuthOption(scenario.authentication, "Folder");
-
-    if (authOption && authOption.scriptLocation === "Folder") {
-      this.scenarioFolder.auth = new RequestAuth({
-        type: "bearer",
-        bearer: [
-          {
-            key: "token",
-            value: `{{${authOption.tokenName}}}`,
-            type: "string",
-          },
-        ],
-      });
-      preScripts.push(
-        PostmanHelper.generateAuthScript(
-          (scenario.authentication as AADTokenAuthentication).scope!,
-          authOption.tokenName
-        )
-      );
+    if (scenario.authentication) {
+      this.scenarioFolder.auth = this.checkAuthOption(scenario.authentication, preScripts, env);
     }
 
     env.resolve();
@@ -258,7 +263,7 @@ export class PostmanCollectionRunnerClient implements ApiScenarioRunnerClient {
 pm.test("Started TestProxy recording", function() {
     pm.response.to.be.success;
     pm.response.to.have.header("x-recording-id");
-    pm.variables.set("x_recording_id", pm.response.headers.get("x-recording-id"));
+    pm.environment.set("x_recording_id", pm.response.headers.get("x-recording-id"));
 });`
     );
   }
@@ -477,26 +482,8 @@ pm.test("Stopped TestProxy recording", function() {
     // pre scripts
     const preScripts: string[] = [];
 
-    step.authentication = env.resolveObjectValues(step.authentication);
-    const authOption = this.checkAuthOption(step.authentication, "Request");
-
-    if (authOption && authOption.scriptLocation === "Request") {
-      item.request.auth = new RequestAuth({
-        type: "bearer",
-        bearer: [
-          {
-            key: "token",
-            value: `{{${authOption.tokenName}}}`,
-            type: "string",
-          },
-        ],
-      });
-      preScripts.push(
-        PostmanHelper.generateAuthScript(
-          (step.authentication as AADTokenAuthentication).scope!,
-          authOption.tokenName
-        )
-      );
+    if (step.authentication) {
+      item.request.auth = this.checkAuthOption(step.authentication, preScripts, env);
     }
 
     item.description = step.operationId;
@@ -618,7 +605,7 @@ pm.test("Stopped TestProxy recording", function() {
       step.responseAssertion
     ).forEach((s) => postScripts.push(s));
 
-    if (step.operation && step.operation["x-ms-long-running-operation"]) {
+    if (step.operation && step.operation[xmsLongRunningOperation]) {
       const metadata: LroItemMetadata = {
         type: "LRO",
         poller_item_name: `_${item.name}_poller`,
@@ -677,7 +664,7 @@ pm.test("Stopped TestProxy recording", function() {
 
     postScripts.push(
       `
-const pollingUrl = pm.response.headers.get("Location") || pm.response.headers.get("Azure-AsyncOperation");
+const pollingUrl = pm.response.headers.get("Location") || pm.response.headers.get("Azure-AsyncOperation") || pm.response.headers.get("Operation-Location");
 if (pollingUrl) {
     pm.variables.set("x_polling_url", ${
       this.opts.testProxy
@@ -731,9 +718,9 @@ try {
             pm.variables.set("x_retry_after", pm.response.headers.get("Retry-After"));
         }
     } else if (pm.response.size().body > 0) {
-        const terminalStatus = ["Succeeded", "Failed", "Canceled"];
+        const terminalStatus = ["succeeded", "failed", "canceled", "cancelled", "aborted", "deleted", "completed"];
         const json = pm.response.json();
-        if (json.status !== undefined && terminalStatus.indexOf(json.status) === -1) {
+        if (json.status !== undefined && terminalStatus.indexOf(json.status.toLowerCase()) === -1) {
             postman.setNextRequest("${delayItem.name}")
             if (pm.response.headers.has("Retry-After")) {
                 pm.variables.set("x_retry_after", pm.response.headers.get("Retry-After"));